split VideoPreview component
separate the images into a component in their own right

this will make adding the alt text field for the trail image easier
